# lib/api-zod/src/generated/types/referralEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneStore Apps Affiliate - Universal Multi-PWA Affiliate Engine
 * OpenAPI spec version: 0.1.0
 */
import type { ReferralEventEventType } from './referralEventEventType';

export interface ReferralEvent {
  id: number;
  affiliateId: number;
  refCode: string;
  appName: string;
  eventType: ReferralEventEventType;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  createdAt: Date;
}