# lib/api-zod/src/generated/types/systemConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneStore Apps Affiliate - Universal Multi-PWA Affiliate Engine
 * OpenAPI spec version: 0.1.0
 */
import type { SystemConfigCommissionType } from './systemConfigCommissionType';

export interface SystemConfig {
  commissionType: SystemConfigCommissionType;
  commissionValue: number;
  holdDays: number;
  apiKey: string;
}