# lib/api-zod/src/generated/types/trackResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OneStore Apps Affiliate - Universal Multi-PWA Affiliate Engine
 * OpenAPI spec version: 0.1.0
 */

export interface TrackResponse {
  success: boolean;
  message: string;
  /** @nullable */
  eventId?: number | null;
}